some changes

+++ C/homework-2/task7.c
@@ -10,29 +10,29 @@ int power(int a, int b){
 
 int main(){
 	
-	int num = 0;
+	long num = 0;
 
 	printf("Enter the number: ");
-	scanf("%d", &num);
+	scanf("%ld", &num);
 	
-	int test = num;
+	long test = num;
 	int count = 0;	
 	while(test > 0){
 		count++;
 		test /= 10;
 	}
 	
-	int n = power(10, count-1);
+	long n = power(10, count-1);
 
 	for(int i = 0; i < count; ++i){
-		int res = (num / n) % 10;
+		long res = (num / n) % 10;
 		if (i == count-1){
-			printf("%d", res);
+			printf("%ld", res);
 		}else{
-			printf("%d, ", res);
-			n/=10;
+			printf("%ld, ", res);
+			n /= 10;
 		}
 	}printf("\n");
-	return 0;
-	
+
+	return 0;	
 }
